Module interface changes.
Modules are now allowed to export a NULL pointer for the function
m_free, in this case, no free is performed on strings returned by m_run.
The type of m_init changed to allow to properly return an error instead
of using a NULL pointer.

Xbar now catches SIGINT and SIGTERM to gracefully release X resources.