Add a silly time module.
This module will simply display the current date and time information,
it is not yet customizable. The static function mod_time_time converts a
struct tm into a string "a la" asctime, however, it does not display
seconds (because, for the sake of simplicity, the refresh rate is not
precise enough).