Cleanup in mod_cpu.
The name CpuLoad was ill-chosen because information in this struct is
not directly related to the load. It was renamed by a more generic name:
CpuInfo.
A static variable last_load was added to avoid dislplaying "Updating..."
when two calls to mod_cpu_run are so close that the cpu info did not
change.