Wait a constant time at each main loop iteration.
The main loop wait was interrupted by X events to redraw the bar, and,
at each redraw, a whole iteration was relaunched. This costed a lot of
cpu time because, modules had to redo their work more often.
This was changed to be sure we wait at least for refresh_time (whose
type moved from int to unsigned) before starting a new main iteration.