Full rewrite of the main event loop.
Modules marked as TRIG_FD now run directly if some data is received on
one of their fds.

This implies a change in data structures because it is no longer
possible to loop over the whole list of modules then filter them
according to their m_trigger falgs.

I used two chained lists, one for TRIG_TIME modules and one for TRIG_FD
modules (one module can be in both, this is not a problem). These two
lists are used in the main loop in mloop to refresh module strings when
needed.

Modules' initialization is now done in a dedicated function (minit).