Parse module strings only when they are updated.
Module strings were parsed at every screen refresh. With this
modification, parsing is done once for all when module strings are
updated.

This adds a lot of code and I am not very satisfied with it. Moreover,
it seems that there is no runtime gain. This will probably stay in this
branch.