Remove config.h and replace it by a default config.
The file config.h is specific to each user. With the default
distribution comes a file config.h.def which will serve as an initial
configuration.

Once the user gets the xbar distribution, it has to run 'make config.h'
to create a default configuration. Then it can customize it and run
'make' to create its own version of xbar.