Handle module strings in the ModRuntime structure.
Module strings were handled locally in mloop in a special array, since
all runtime information of modules was stored in a specific ModRuntime
struct, it is more natural to store these strings (which are runtime
iformation) in this data structure. This change also saves a copy of the
pack modes in a special array.

The action of running a module was abstracted in a new mrun function.